@@ -1,6 +1,6 @@
-import 'package:flutter/cupertino.dart';
+import 'package:flutter/foundation.dart';
 
-class Product{
+class Product {
   final String id;
   final String title;
   final String description;
@@ -9,11 +9,11 @@ class Product{
   bool isFavorite;
 
   Product({
-   @required this.id,
-  @required this.title,
-  @required this.description,
-  @required this.price,
-  @required this.imageUrl,
-  this.isFavorite = false
-      });
+    @required this.id,
+    @required this.title,
+    @required this.description,
+    @required this.price,
+    @required this.imageUrl,
+    this.isFavorite = false,
+  });
 }
